fix(workflow): migrate to gtexquery v1.0.0
By moving to GTExQuery v1.0.0 and moving to a csv output, we eliminate the testing inconsitencies of excel files.

BREAKING CHANGE: The output is now a collection of csv files, one per gene. As mentioned above, this eliminates the need to try to test excel file. It also greatly simplifies multithreading. Closes #1.