feat(workflow): provides resources in repository
To simplify running the pipeline, the gencode annotations and MANE summary are provided within the repository as resources. This simplifies testing (fewer steps, and symlinks mean only one copy of the data is included) and removes 2 API queries, improving reproducibility.

BREAKING CHANGE: While the output of the pipeline remains unchanged, users who do not already have the resources locally must either generate them or pull them from the repository. As this changes the input requirements of the pipeline, I have decided to class this as Breaking.